separated Vcf and Bedpe classes into 2 files.
