+++ tests/ml/data/example_signature_definition.txt
@@ -2,40 +2,40 @@
 # example file of custom SV class definitione
 
 name "small deletion"
-1 SVTYPE == DEL
-2 SVLEN > -1000000
-3 SVLEN < 0
-logic 1 & 2 & 3 
+0 SVTYPE == DEL
+1 SVLEN > -1000000
+2 SVLEN < 0
+logic 0 & 1 & 2 
 
 name "large deletion"
-1 SVTYPE == DEL
-2 SVLEN < -1000000
-logic 1 & 2
+0 SVTYPE == DEL
+1 SVLEN < -1000000
+logic 0 & 1
 
 name "small duplication"
-1 SVTYPE == DUP
-2 SVLEN > -1000000
-3 SVLEN < 0
-logic 1 & 2 & 3 
+0 SVTYPE == DUP
+1 SVLEN > -1000000
+2 SVLEN < 0
+logic 0 & 1 & 2 
 
 name "large duplication"
-1 SVTYPE == DUP
-2 SVLEN < -1000000
-logic 1 & 2
+0 SVTYPE == DUP
+1 SVLEN < -1000000
+logic 0 & 1
 
 name "small inversion"
-1 SVTYPE == t2tINV
-2 SVTYPE == h2hINV
-3 SVLEN > -1000000
-4 SVLEN < 0
-logic (1 | 2) & 3 & 4
+0 SVTYPE == t2tINV
+1 SVTYPE == h2hINV
+2 SVLEN > -1000000
+3 SVLEN < 0
+logic (0 | 1) & 2 & 3
 
 name "large inversion"
-1 SVTYPE == t2tINV
-2 SVTYPE == h2hINV
+0 SVTYPE == t2tINV
+1 SVTYPE == h2hINV
 2 SVLEN < -1000000
-logic (1 | 2) & 3
+logic (0 | 1) & 2
 
 name "translocation"
-1 SVTYPE == TRA
-logic 1
+0 SVTYPE == TRA
+logic 0
